loss does not go down
