+++ configs/walking_rnn/MB_ft_walking_rnn.yaml
@@ -1,0 +1,33 @@
+# General
+finetune: True
+partial_train: null
+
+# Traning
+epochs: 20
+batch_size: 16
+lr_backbone: 0.0001
+lr_head: 0.00003
+weight_decay: 0.03
+lr_decay: 0.99
+
+# Model
+maxlen: 243
+dim_feat: 512
+mlp_ratio: 2
+depth: 5
+dim_rep: 512
+num_heads: 8
+att_fuse: True
+num_joints: 17
+hidden_dim: 2048
+dropout_ratio: 0.3
+
+# Data
+dataset: walking
+data_split: xsub
+clip_len: 243
+
+# Augmentation
+random_move: True
+scale_range_train: [1, 3]
+scale_range_test: [2, 2]
